+++ app/src/app/examples/calendar/calendar.config.js
@@ -57,7 +57,7 @@
             // set the menu type to a link
             type: 'link',
             // set an icon for this menu
-            icon: 'icon-today',
+            icon: 'zmdi zmdi-calendar-alt',
             // set a proirity for this menu item, menu is sorted by priority
             priority: 2.3
         });