+++ app/src/app/examples/dashboards/server/dashboard-server.controller.js
@@ -6,18 +6,18 @@
         .controller('DashboardServerController', DashboardServerController);
 
     /* @ngInject */
-    function DashboardServerController($scope, $timeout, $mdToast) {
+    function DashboardServerController($scope, $timeout, $mdToast, $element, dragulaService) {
         var vm = this;
         vm.disks = [{
-            icon: 'icon-storage',
+            icon: 'zmdi zmdi-storage',
             name: 'Ubuntu 10.04 LTS Disk Image (10000 MB, ext3)',
             enabled: true
         },{
-            icon: 'icon-settings-input-component',
+            icon: 'zmdi zmdi-settings-input-component',
             name: 'Ubuntu 11.10 SSD Image (10224 MB, ext3)',
             enabled: false
         },{
-            icon: 'icon-storage',
+            icon: 'zmdi zmdi-storage',
             name: '256MB Swap Image (256 MB, swap)',
             enabled: true
         }];
@@ -45,5 +45,10 @@
                 .hideDelay(3000)
             );
         }, 5000);
+
+        var mirrorContainer = $element.find('.mirror-container')[0];
+        dragulaService.options($scope,'drag-server-container', {
+            mirrorContainer: mirrorContainer
+        });
     }
 })();